Run an UDF on a `stars` object created from generic GeoTIFF files

Reads generic files from disk by looking up a legend file, runs the User Defined Function (UDF) specified by the user and writes back the results to disk as a multi-band GeoTIFF file named `out.tif` in the directory specified by the user.

run_UDF(legend_name = "legend.csv", function_name, drop_dim,
  in_dim = c(1, 1, 1, 1, 1), out_dir = "results")

Arguments

legend_name
Name of the legend file as a string (default: "legend.csv")

function_name
Name of the User-Defined Function (UDF) the user defined in his own script

drop_dim
Numeric value (or vector) representing the dimension number of the dimension to be dropped. (1,2 = Space, 3 = Band, 4 = Time, 5 = Whether raster)

in_dim
Dimensionality of the incoming Collection object (default: `c(1,1,1,1,1)` representing spatial multi-band, multi-temporal raster)

out_dir
Path of the directory where the resulting files are to be written to disk (default: "results")

Details

The semantics of the written multi-band GeoTIFF depends on the argument `drop_dim`. For example, if `drop_dim = 4` the bands in the output file represent bands while if `drop_dim = 3`, the bands in the output file represent time. This modification of the dimensionality (`drop_dim`) as well as the dimensionality of the incoming Collection object (`in_dim`) would be controlled by the backend as parameters passed on to the UDF server as metadata files along with the legend file (`legend_name`).